refactor: Board 엔티티 삭제

+++ src/main/java/com/soundbar91/retrospect_project/service/PostService.java
@@ -3,13 +3,13 @@ package com.soundbar91.retrospect_project.service;
 import com.soundbar91.retrospect_project.controller.dto.request.RequestCreatePost;
 import com.soundbar91.retrospect_project.controller.dto.request.RequestUpdatePost;
 import com.soundbar91.retrospect_project.controller.dto.response.ResponsePost;
-import com.soundbar91.retrospect_project.entity.Board;
 import com.soundbar91.retrospect_project.entity.Post;
+import com.soundbar91.retrospect_project.entity.Problem;
 import com.soundbar91.retrospect_project.entity.User;
 import com.soundbar91.retrospect_project.entity.keyInstance.Category;
 import com.soundbar91.retrospect_project.exception.ApplicationException;
-import com.soundbar91.retrospect_project.repository.BoardRepository;
 import com.soundbar91.retrospect_project.repository.PostRepository;
+import com.soundbar91.retrospect_project.repository.ProblemRepository;
 import com.soundbar91.retrospect_project.repository.UserRepository;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.TypedQuery;
@@ -21,8 +21,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.soundbar91.retrospect_project.exception.errorCode.BoardErrorCode.NOT_FOUND_BOARD;
 import static com.soundbar91.retrospect_project.exception.errorCode.PostErrorCode.NOT_FOUND_POST;
+import static com.soundbar91.retrospect_project.exception.errorCode.ProblemErrorCode.NOT_FOUND_PROBLEM;
 import static com.soundbar91.retrospect_project.exception.errorCode.UserErrorCode.NOT_FOUND_USER;
 import static com.soundbar91.retrospect_project.exception.errorCode.UserErrorCode.NOT_PERMISSION;
 
@@ -32,45 +32,37 @@ public class PostService {
 
     private final EntityManager entityManager;
     private final PostRepository postRepository;
-    private final BoardRepository boardRepository;
+    private final ProblemRepository problemRepository;
     private final UserRepository userRepository;
 
     @Transactional
     public void createPost(
             RequestCreatePost requestCreatePost,
-            Long boardId,
+            Long problemId,
             HttpServletRequest httpServletRequest
     ) {
         Long userId = (Long) httpServletRequest.getSession().getAttribute("userId");
         User user = userRepository.findById(userId)
                 .orElseThrow(() -> new ApplicationException(NOT_FOUND_USER));
-        Board board = boardRepository.findById(boardId)
-                .orElseThrow(() -> new ApplicationException(NOT_FOUND_BOARD));
+        Problem problem = problemRepository.findById(problemId)
+                .orElseThrow(() -> new ApplicationException(NOT_FOUND_PROBLEM));
 
-        postRepository.save(requestCreatePost.toEntity(user, board));
+        postRepository.save(requestCreatePost.toEntity(user, problem));
     }
 
     public List<ResponsePost> getPosts(
-            Long boardId, String username, Category category
+            Long problemId, String username, Category category
     ) {
-        Board board = null;
-        if (boardId != null) {
-            board = boardRepository.findById(boardId)
-                    .orElseThrow(() -> new ApplicationException(NOT_FOUND_BOARD));
-        }
+        StringBuilder jpql = createJpql(problemId, username, category);
 
-        User user = null;
-        if (username != null) {
-            user = userRepository.findByUsername(username)
-                    .orElseThrow(() -> new ApplicationException(NOT_FOUND_USER));
-        }
+        Problem problem = problemId != null ?
+                problemRepository.findById(problemId).orElseThrow(() -> new ApplicationException(NOT_FOUND_PROBLEM)) : null;
 
-        StringBuilder jpql = getJpql(board, user, category);
+        User user = username != null ?
+                userRepository.findByUsername(username).orElseThrow(() -> new ApplicationException(NOT_FOUND_USER)) : null;
 
         TypedQuery<Post> query = entityManager.createQuery(jpql.toString(), Post.class);
-        if (user != null) query.setParameter("user", user);
-        if (board != null) query.setParameter("board", board);
-        if (category != null) query.setParameter("category", category);
+        queryParameterBinding(category, user, query, problem);
 
         return query.getResultList().stream().map(ResponsePost::from).toList();
     }
@@ -87,7 +79,7 @@ public class PostService {
             HttpServletRequest httpServletRequest,
             Long postId
     ) {
-        Post post = valid(postId, httpServletRequest);
+        Post post = checkPermission(postId, httpServletRequest);
         post.updatePost(requestUpdatePost);
         postRepository.flush();
     }
@@ -102,16 +94,16 @@ public class PostService {
 
     @Transactional
     public void deletePost(Long postId, HttpServletRequest httpServletRequest) {
-        valid(postId, httpServletRequest);
+        checkPermission(postId, httpServletRequest);
         postRepository.deleteById(postId);
     }
 
-    private StringBuilder getJpql(Board board, User user, Category category) {
+    private StringBuilder createJpql(Long problemId, String username, Category category) {
         StringBuilder jpql = new StringBuilder("select p from Post p");
         List<String> criteria = new ArrayList<>();
 
-        if (board != null) criteria.add(" p.board = :board");
-        if (user != null) criteria.add(" p.user = :user");
+        if (problemId != null) criteria.add(" p.problem = :problem");
+        if (username != null) criteria.add(" p.user = :user");
         if (category != null) criteria.add(" p.category = :category");
         if (!criteria.isEmpty()) jpql.append(" where ");
 
@@ -123,7 +115,13 @@ public class PostService {
         return jpql;
     }
 
-    private Post valid(Long postId, HttpServletRequest httpServletRequest) {
+    private void queryParameterBinding(Category category, User user, TypedQuery<Post> query, Problem problem) {
+        if (user != null) query.setParameter("user", user);
+        if (problem != null) query.setParameter("problem", problem);
+        if (category != null) query.setParameter("category", category);
+    }
+
+    private Post checkPermission(Long postId, HttpServletRequest httpServletRequest) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new ApplicationException(NOT_FOUND_POST));
